requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.10.4
firebase-admin==6.6.0
